+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2023-06-26
+### Changed
+- [@mjkuranda](https://github.com/mjkuranda): GameMap to GameData.
+- [@mjkuranda](https://github.com/mjkuranda): Spawn and destroy refactored.
+
+### Fixed
+- [@mjkuranda](https://github.com/mjkuranda): ArrayOutOfBoundsException related to player movement and missiles.
+- [@mjkuranda](https://github.com/mjkuranda): Building artifacts again.
+
 ## [0.4.0] - 2023-06-23
 ### Added
 - [@mjkuranda](https://github.com/mjkuranda): GameMap and Renderer abstract classes.